Extend precision of ADC Using a u16 to give a wide range. so implementations of the trait can use (truncate to) the appropriate range desired

+++ baseline-sim/src/sim.rs
@@ -617,7 +617,7 @@ where
         Adc::get_states(self.interp.get_peripherals())
     }
 
-    fn get_adc_readings(&self) -> AdcPinArr<Result<u8, AdcReadError>> {
+    fn get_adc_readings(&self) -> AdcPinArr<Result<u16, AdcReadError>> {
         Adc::read_all(self.interp.get_peripherals())
     }
