+++ traits/src/control/rpc/controller.rs
@@ -480,7 +480,7 @@ where
     fn get_gpio_states(&self) -> GpioPinArr<GpioState> { ctrl!(self, GetGpioStates, R::GetGpioStates(r), r) }
     fn get_gpio_readings(&self) -> GpioPinArr<Result<bool, GpioReadError>> { ctrl!(self, GetGpioReadings, R::GetGpioReadings(r), r) }
     fn get_adc_states(&self) -> AdcPinArr<AdcState> { ctrl!(self, GetAdcStates, R::GetAdcStates(r), r) }
-    fn get_adc_readings(&self) -> AdcPinArr<Result<u8, AdcReadError>> { ctrl!(self, GetAdcReadings, R::GetAdcReadings(r), r) }
+    fn get_adc_readings(&self) -> AdcPinArr<Result<u16, AdcReadError>> { ctrl!(self, GetAdcReadings, R::GetAdcReadings(r), r) }
     fn get_timer_modes(&self) -> TimerArr<TimerMode> { ctrl!(self, GetTimerModes, R::GetTimerModes(r), r) }
     fn get_timer_states(&self) -> TimerArr<TimerState> { ctrl!(self, GetTimerStates, R::GetTimerStates(r), r) }
     fn get_pwm_states(&self) -> PwmPinArr<PwmState> { ctrl!(self, GetPwmStates, R::GetPwmStates(r), r) }
