+++ traits/src/peripherals/stubs.rs
@@ -41,7 +41,7 @@ impl Adc for AdcStub {
     fn set_state(&mut self, _pin: AdcPin, _: AdcState) -> Result<(), AdcMiscError> { Err(AdcMiscError) }
     fn get_state(&self, _pin: AdcPin) -> AdcState { AdcState::Disabled }
 
-    fn read(&self, pin: AdcPin) -> Result<u8, AdcReadError> { Err(AdcReadError((pin, AdcState::Disabled)))}
+    fn read(&self, pin: AdcPin) -> Result<u16, AdcReadError> { Err(AdcReadError((pin, AdcState::Disabled)))}
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Default)]
